Fixed issue caused by sideefect of lazy published repo
