+++ src/main/java/org/craftercms/studio/impl/v1/repository/git/GitContentRepository.java
@@ -78,6 +78,7 @@ import org.craftercms.studio.api.v1.service.security.SecurityService;
 import org.craftercms.studio.api.v1.to.RemoteRepositoryInfoTO;
 import org.craftercms.studio.api.v1.to.VersionTO;
 import org.craftercms.studio.api.v2.service.security.internal.UserServiceInternal;
+import org.craftercms.studio.api.v2.utils.GitRepositoryHelper;
 import org.craftercms.studio.api.v2.utils.StudioConfiguration;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.LsRemoteCommand;
@@ -170,7 +171,6 @@ import static org.eclipse.jgit.transport.RemoteRefUpdate.Status.REJECTED_REMOTE_
 public class GitContentRepository implements ContentRepository, ServletContextAware {
 
     private static final Logger logger = LoggerFactory.getLogger(GitContentRepository.class);
-    private GitContentRepositoryHelper helper = null;
     private TextEncryptor encryptor;
 
     private static final String STUDIO_MANIFEST_LOCATION = "/META-INF/MANIFEST.MF";
@@ -187,9 +187,11 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public boolean contentExists(String site, String path) {
         boolean toReturn = false;
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-        if (repo != null ) {
-            try {
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+            if (repo != null) {
                 RevTree tree = helper.getTreeForLastCommit(repo);
                 try (TreeWalk tw = TreeWalk.forPath(repo, helper.getGitPath(path), tree)) {
                     // Check if the array of items is not null, and since we have an absolute path to the item,
@@ -206,9 +208,9 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
                 } catch (IOException e) {
                     logger.info("Content not found for site: " + site + " path: " + path, e);
                 }
-            } catch (Exception e) {
-                logger.error("Failed to create RevTree for site: " + site + " path: " + path, e);
             }
+        } catch (Exception e) {
+            logger.error("Failed to create RevTree for site: " + site + " path: " + path, e);
         }
         return toReturn;
     }
@@ -216,12 +218,13 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public InputStream getContent(String site, String path) throws ContentNotFoundException {
         InputStream toReturn = null;
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-        if (repo == null) {
-            throw new ContentNotFoundException("Repository not found for site " + site);
-        }
-
         try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+            if (repo == null) {
+                throw new ContentNotFoundException("Repository not found for site " + site);
+            }
             RevTree tree = helper.getTreeForLastCommit(repo);
             try (TreeWalk tw = TreeWalk.forPath(repo, helper.getGitPath(path), tree)) {
                 // Check if the array of items is not null, and since we have an absolute path to the item,
@@ -235,7 +238,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
             } catch (IOException e) {
                 logger.error("Error while getting content for file at site: " + site + " path: " + path, e);
             }
-        } catch (IOException e) {
+        } catch (IOException | CryptoException e) {
             logger.error("Failed to create RevTree for site: " + site + " path: " + path, e);
         }
 
@@ -244,8 +247,10 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public long getContentSize(final String site, final String path) {
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
         try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
             RevTree tree = helper.getTreeForLastCommit(repo);
             try (TreeWalk tw = TreeWalk.forPath(repo, helper.getGitPath(path), tree)) {
                 if (tw != null && tw.getObjectId(0) != null) {
@@ -254,7 +259,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
                     return objectLoader.getSize();
                 }
             }
-        } catch (IOException e) {
+        } catch (IOException | CryptoException e) {
             logger.error("Error while getting content for file at site: " + site + " path: " + path, e);
         }
         return -1L;
@@ -264,30 +269,33 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     public String writeContent(String site, String path, InputStream content) {
         // Write content to git and commit it
         String commitId = null;
-
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Repository repo =
-                helper.getRepository(site, StringUtils.isEmpty(site)? GLOBAL: SANDBOX);
-
-            try {
-
-                if (repo != null) {
-                    if (helper.writeFile(repo, site, path, content)) {
-                        PersonIdent user = helper.getCurrentUserIdent();
-                        String username = securityService.getCurrentUser();
-                        String comment = helper.getCommitMessage(REPO_SANDBOX_WRITE_COMMIT_MESSAGE)
-                            .replace(REPO_COMMIT_MESSAGE_USERNAME_VAR, username)
-                            .replace(REPO_COMMIT_MESSAGE_PATH_VAR, path);
-                        commitId = helper.commitFile(repo, site, path, comment, user);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site)? GLOBAL: SANDBOX);
+                try {
+                    if (repo != null) {
+                        if (helper.writeFile(repo, site, path, content)) {
+                            PersonIdent user = helper.getCurrentUserIdent();
+                            String username = securityService.getCurrentUser();
+                            String comment = helper.getCommitMessage(REPO_SANDBOX_WRITE_COMMIT_MESSAGE)
+                                    .replace(REPO_COMMIT_MESSAGE_USERNAME_VAR, username)
+                                    .replace(REPO_COMMIT_MESSAGE_PATH_VAR, path);
+                            commitId = helper.commitFile(repo, site, path, comment, user);
+                        } else {
+                            logger.error("Failed to write content site: " + site + " path: " + path);
+                        }
                     } else {
-                        logger.error("Failed to write content site: " + site + " path: " + path);
+                        logger.error("Missing repository during write for site: " + site + " path: " + path);
                     }
-                } else {
-                    logger.error("Missing repository during write for site: " + site + " path: " + path);
+                }  catch (ServiceLayerException | UserNotFoundException e) {
+                    logger.error("Unknown service error during write for site: " + site + " path: " + path, e);
                 }
-            } catch (ServiceLayerException | UserNotFoundException e) {
-                logger.error("Unknown service error during write for site: " + site + " path: " + path, e);
+
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected service error during write for site: " + site + " path: " + path, e);
         }
 
         return commitId;
@@ -298,98 +306,110 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
         // SJ: Git doesn't care about empty folders, so we will create the folders and put a 0 byte file in them
         String commitId = null;
         boolean result;
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Path emptyFilePath = Paths.get(path, name, EMPTY_FILE);
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Path emptyFilePath = Paths.get(path, name, EMPTY_FILE);
-            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-
-            try {
-                // Create basic file
-                File file = new File(repo.getDirectory().getParent(), emptyFilePath.toString());
-
-                // Create parent folders
-                File folder = file.getParentFile();
-                if (folder != null) {
-                    if (!folder.exists()) {
-                        folder.mkdirs();
-                    }
-                }
-
-                // Create the file
-                if (!file.createNewFile()) {
-                    logger.error("error writing file: site: " + site + " path: " + emptyFilePath);
-                    result = false;
-                } else {
-                    // Add the file to git
-                    try (Git git = new Git(repo)) {
-                        git.add().addFilepattern(helper.getGitPath(emptyFilePath.toString())).call();
-
-                        git.close();
-                        result = true;
-                    } catch (GitAPIException e) {
-                        logger.error("error adding file to git: site: " + site + " path: " + emptyFilePath, e);
-                        result = false;
-                    }
-                }
-            } catch (IOException e) {
-                logger.error("error writing file: site: " + site + " path: " + emptyFilePath, e);
-                result = false;
-            }
-
-            if (result) {
                 try {
-                    commitId = helper.commitFile(repo, site, emptyFilePath.toString(),
-                            helper.getCommitMessage(REPO_CREATE_FOLDER_COMMIT_MESSAGE)
-                                    .replaceAll(PATTERN_SITE, site)
-                                    .replaceAll(PATTERN_PATH, path + FILE_SEPARATOR + name),
-                            helper.getCurrentUserIdent());
-                } catch (ServiceLayerException | UserNotFoundException e) {
-                    logger.error("Unknown service error during commit for site: " + site + " path: "
-                        + emptyFilePath, e);
+                    // Create basic file
+                    File file = new File(repo.getDirectory().getParent(), emptyFilePath.toString());
+
+                    // Create parent folders
+                    File folder = file.getParentFile();
+                    if (folder != null) {
+                        if (!folder.exists()) {
+                            folder.mkdirs();
+                        }
+                    }
+
+                    // Create the file
+                    if (!file.createNewFile()) {
+                        logger.error("error writing file: site: " + site + " path: " + emptyFilePath);
+                        result = false;
+                    } else {
+                        // Add the file to git
+                        try (Git git = new Git(repo)) {
+                            git.add().addFilepattern(helper.getGitPath(emptyFilePath.toString())).call();
+
+                            git.close();
+                            result = true;
+                        } catch (GitAPIException e) {
+                            logger.error("error adding file to git: site: " + site + " path: " + emptyFilePath, e);
+                            result = false;
+                        }
+                    }
+                } catch (IOException e) {
+                    logger.error("error writing file: site: " + site + " path: " + emptyFilePath, e);
+                    result = false;
+                }
+
+                if (result) {
+                    try {
+                        commitId = helper.commitFile(repo, site, emptyFilePath.toString(),
+                                helper.getCommitMessage(REPO_CREATE_FOLDER_COMMIT_MESSAGE)
+                                        .replaceAll(PATTERN_SITE, site)
+                                        .replaceAll(PATTERN_PATH, path + FILE_SEPARATOR + name),
+                                helper.getCurrentUserIdent());
+                    } catch (ServiceLayerException | UserNotFoundException e) {
+                        logger.error("Unknown service error during commit for site: " + site + " path: "
+                                + emptyFilePath, e);
+                    }
                 }
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error writing file: site: " + site + " path: " + path, e);
+            result = false;
         }
-
         return commitId;
     }
 
     @Override
     public String deleteContent(String site, String path, String approver) {
         String commitId = null;
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+                try (Git git = new Git(repo)) {
+                    String pathToDelete = helper.getGitPath(path);
+                    Path toDelete = Paths.get(repo.getDirectory().getParent(), pathToDelete);
+                    Path parentToDelete = Paths.get(pathToDelete).getParent();
+                    git.rm().addFilepattern(pathToDelete).setCached(false).call();
 
-            try (Git git = new Git(repo)) {
-                String pathToDelete = helper.getGitPath(path);
-                Path toDelete = Paths.get(repo.getDirectory().getParent(), pathToDelete);
-                Path parentToDelete = Paths.get(pathToDelete).getParent();
-                git.rm().addFilepattern(pathToDelete).setCached(false).call();
+                    String pathToCommit = pathToDelete;
+                    if (toDelete.toFile().isFile()) {
+                        pathToCommit = deleteParentFolder(git, parentToDelete);
+                    }
 
-                String pathToCommit = pathToDelete;
-                if (toDelete.toFile().isFile()) {
-                    pathToCommit = deleteParentFolder(git, parentToDelete);
+                    // TODO: SJ: we need to define messages in a string table of sorts
+                    commitId = helper.commitFile(repo, site, pathToCommit,
+                            helper.getCommitMessage(REPO_DELETE_CONTENT_COMMIT_MESSAGE)
+                                    .replaceAll(PATTERN_PATH, path),
+                            StringUtils.isEmpty(approver) ? helper.getCurrentUserIdent() : helper.getAuthorIdent(approver));
+
+                } catch (GitAPIException | UserNotFoundException e) {
+                    logger.error("Error while deleting content for site: " + site + " path: " + path, e);
+                } catch (ServiceLayerException e) {
+                    logger.error("Unknown service error during delete for site: " + site + " path: " + path, e);
                 }
-
-                // TODO: SJ: we need to define messages in a string table of sorts
-                commitId = helper.commitFile(repo, site, pathToCommit,
-                        helper.getCommitMessage(REPO_DELETE_CONTENT_COMMIT_MESSAGE)
-                                .replaceAll(PATTERN_PATH, path),
-                        StringUtils.isEmpty(approver) ? helper.getCurrentUserIdent() : helper.getAuthorIdent(approver));
-
-            } catch (GitAPIException | UserNotFoundException e) {
-                logger.error("Error while deleting content for site: " + site + " path: " + path, e);
-            } catch (ServiceLayerException e) {
-                logger.error("Unknown service error during delete for site: " + site + " path: " + path, e);
             }
+        }  catch (CryptoException e) {
+            logger.error("Unexpected service error during delete for site: " + site + " path: " + path, e);
         }
 
         return commitId;
     }
 
-    private String deleteParentFolder(Git git, Path parentFolder) throws GitAPIException {
+    private String deleteParentFolder(Git git, Path parentFolder) throws GitAPIException, CryptoException {
         String parent = parentFolder.toString();
         String toRet = parent;
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
         String folderToDelete = helper.getGitPath(parent);
         Path toDelete = Paths.get(git.getRepository().getDirectory().getParent(), parent);
         String[] children = toDelete.toFile().list();
@@ -415,122 +435,133 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     public Map<String, String> moveContent(String site, String fromPath, String toPath, String newName) {
         Map<String, String> toRet = new TreeMap<String, String>();
         String commitId;
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-            String gitFromPath = helper.getGitPath(fromPath);
-            String gitToPath;
-            if (StringUtils.isEmpty(newName)) {
-                gitToPath = helper.getGitPath(toPath);
-            } else {
-                gitToPath = helper.getGitPath(toPath + FILE_SEPARATOR + newName);
-            }
-
-            try (Git git = new Git(repo)) {
-                // Check if destination is a file, then this is a rename operation
-                // Perform rename and exit
-                Path sourcePath = Paths.get(repo.getDirectory().getParent(), gitFromPath);
-                File sourceFile = sourcePath.toFile();
-                Path targetPath = Paths.get(repo.getDirectory().getParent(), gitToPath);
-                File targetFile = targetPath.toFile();
-
-                if (sourceFile.getCanonicalFile().equals(targetFile.getCanonicalFile())) {
-                    sourceFile.renameTo(targetFile);
+                String gitFromPath = helper.getGitPath(fromPath);
+                String gitToPath;
+                if (StringUtils.isEmpty(newName)) {
+                    gitToPath = helper.getGitPath(toPath);
                 } else {
-                    if (targetFile.isFile()) {
-                        if (sourceFile.isFile()) {
-                            sourceFile.renameTo(targetFile);
+                    gitToPath = helper.getGitPath(toPath + FILE_SEPARATOR + newName);
+                }
+
+                try (Git git = new Git(repo)) {
+                    // Check if destination is a file, then this is a rename operation
+                    // Perform rename and exit
+                    Path sourcePath = Paths.get(repo.getDirectory().getParent(), gitFromPath);
+                    File sourceFile = sourcePath.toFile();
+                    Path targetPath = Paths.get(repo.getDirectory().getParent(), gitToPath);
+                    File targetFile = targetPath.toFile();
+
+                    if (sourceFile.getCanonicalFile().equals(targetFile.getCanonicalFile())) {
+                        sourceFile.renameTo(targetFile);
+                    } else {
+                        if (targetFile.isFile()) {
+                            if (sourceFile.isFile()) {
+                                sourceFile.renameTo(targetFile);
+                            } else {
+                                // This is not a valid operation
+                                logger.error("Invalid move operation: Trying to rename a directory to a file " +
+                                        "for site: " + site + " fromPath: " + fromPath + " toPath: " + toPath +
+                                        " newName: " + newName);
+                            }
+                        } else if (sourceFile.isDirectory()) {
+                            // Check if we're moving a single file or whole subtree
+                            File[] dirList = sourceFile.listFiles();
+                            for (File child : dirList) {
+                                if (!child.equals(sourceFile)) {
+                                    FileUtils.moveToDirectory(child, targetFile, true);
+                                }
+                            }
+                            FileUtils.deleteDirectory(sourceFile);
                         } else {
-                            // This is not a valid operation
-                            logger.error("Invalid move operation: Trying to rename a directory to a file " +
-                                    "for site: " + site + " fromPath: " + fromPath + " toPath: " + toPath +
-                                    " newName: " + newName);
-                        }
-                    } else if (sourceFile.isDirectory()) {
-                        // Check if we're moving a single file or whole subtree
-                        File[] dirList = sourceFile.listFiles();
-                        for (File child : dirList) {
-                            if (!child.equals(sourceFile)) {
-                                FileUtils.moveToDirectory(child, targetFile, true);
+                            if (sourceFile.isFile()) {
+                                FileUtils.moveFile(sourceFile, targetFile);
+                            } else {
+                                FileUtils.moveToDirectory(sourceFile, targetFile, true);
                             }
                         }
-                        FileUtils.deleteDirectory(sourceFile);
-                    } else {
-                        if (sourceFile.isFile()) {
-                            FileUtils.moveFile(sourceFile, targetFile);
-                        } else {
-                            FileUtils.moveToDirectory(sourceFile, targetFile, true);
-                        }
                     }
+
+                    // The operation is done on disk, now it's time to commit
+                    git.add().addFilepattern(gitToPath).call();
+
+                    Status gitStatus = git.status().addPath(gitToPath).call();
+                    Set<String> changeSet = gitStatus.getAdded();
+
+                    for (String pathToCommit : changeSet) {
+                        String pathRemoved = pathToCommit.replace(gitToPath, gitFromPath);
+                        RevCommit commit = git.commit()
+                                .setOnly(pathToCommit)
+                                .setOnly(pathRemoved)
+                                .setAuthor(helper.getCurrentUserIdent())
+                                .setCommitter(helper.getCurrentUserIdent())
+                                .setMessage(helper.getCommitMessage(REPO_MOVE_CONTENT_COMMIT_MESSAGE)
+                                        .replaceAll(PATTERN_FROM_PATH, fromPath)
+                                        .replaceAll(PATTERN_TO_PATH, toPath +
+                                                (StringUtils.isNotEmpty(newName) ? newName : EMPTY)))
+                                .call();
+                        commitId = commit.getName();
+                        toRet.put(pathToCommit, commitId);
+                    }
+                } catch (IOException | GitAPIException | ServiceLayerException | UserNotFoundException e) {
+                    logger.error("Error while moving content for site: " + site + " fromPath: " + fromPath +
+                            " toPath: " + toPath + " newName: " + newName);
                 }
-
-                // The operation is done on disk, now it's time to commit
-                git.add().addFilepattern(gitToPath).call();
-
-                Status gitStatus = git.status().addPath(gitToPath).call();
-                Set<String> changeSet = gitStatus.getAdded();
-
-                for (String pathToCommit : changeSet) {
-                    String pathRemoved = pathToCommit.replace(gitToPath, gitFromPath);
-                    RevCommit commit = git.commit()
-                            .setOnly(pathToCommit)
-                            .setOnly(pathRemoved)
-                            .setAuthor(helper.getCurrentUserIdent())
-                            .setCommitter(helper.getCurrentUserIdent())
-                            .setMessage(helper.getCommitMessage(REPO_MOVE_CONTENT_COMMIT_MESSAGE)
-                                    .replaceAll(PATTERN_FROM_PATH, fromPath)
-                                    .replaceAll(PATTERN_TO_PATH, toPath +
-                                            (StringUtils.isNotEmpty(newName) ? newName : EMPTY)))
-                            .call();
-                    commitId = commit.getName();
-                    toRet.put(pathToCommit, commitId);
-                }
-            } catch (IOException | GitAPIException | ServiceLayerException | UserNotFoundException e) {
-                logger.error("Error while moving content for site: " + site + " fromPath: " + fromPath +
-                        " toPath: " + toPath + " newName: " + newName);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while moving content for site: " + site + " fromPath: " + fromPath +
+                    " toPath: " + toPath + " newName: " + newName);
         }
-
         return toRet;
     }
 
     @Override
     public String copyContent(String site, String fromPath, String toPath) {
         String commitId = null;
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+                String gitFromPath = helper.getGitPath(fromPath);
+                String gitToPath = helper.getGitPath(toPath);
 
-            String gitFromPath = helper.getGitPath(fromPath);
-            String gitToPath = helper.getGitPath(toPath);
+                try (Git git = new Git(repo)) {
+                    Path sourcePath = Paths.get(repo.getDirectory().getParent(), fromPath);
+                    File sourceFile = sourcePath.toFile();
+                    Path targetPath = Paths.get(repo.getDirectory().getParent(), toPath);
+                    File targetFile = targetPath.toFile();
 
-            try (Git git = new Git(repo)) {
-                Path sourcePath = Paths.get(repo.getDirectory().getParent(), fromPath);
-                File sourceFile = sourcePath.toFile();
-                Path targetPath = Paths.get(repo.getDirectory().getParent(), toPath);
-                File targetFile = targetPath.toFile();
+                    // Check if we're copying a single file or whole subtree
+                    FileUtils.copyDirectory(sourceFile, targetFile);
 
-                // Check if we're copying a single file or whole subtree
-                FileUtils.copyDirectory(sourceFile, targetFile);
+                    // The operation is done on disk, now it's time to commit
+                    git.add().addFilepattern(gitToPath).call();
+                    RevCommit commit = git.commit()
+                            .setOnly(gitFromPath)
+                            .setOnly(gitToPath)
+                            .setAuthor(helper.getCurrentUserIdent())
+                            .setCommitter(helper.getCurrentUserIdent())
+                            .setMessage(helper.getCommitMessage(REPO_COPY_CONTENT_COMMIT_MESSAGE)
+                                    .replaceAll(PATTERN_FROM_PATH, fromPath).replaceAll(PATTERN_TO_PATH, toPath))
+                            .call();
+                    commitId = commit.getName();
 
-                // The operation is done on disk, now it's time to commit
-                git.add().addFilepattern(gitToPath).call();
-                RevCommit commit = git.commit()
-                        .setOnly(gitFromPath)
-                        .setOnly(gitToPath)
-                        .setAuthor(helper.getCurrentUserIdent())
-                        .setCommitter(helper.getCurrentUserIdent())
-                        .setMessage(helper.getCommitMessage(REPO_COPY_CONTENT_COMMIT_MESSAGE)
-                                .replaceAll(PATTERN_FROM_PATH, fromPath).replaceAll(PATTERN_TO_PATH, toPath))
-                        .call();
-                commitId = commit.getName();
-
-            } catch (IOException | GitAPIException | ServiceLayerException | UserNotFoundException e) {
-                logger.error("Error while copying content for site: " + site + " fromPath: " + fromPath +
-                        " toPath: " + toPath + " newName: ");
+                } catch (IOException | GitAPIException | ServiceLayerException | UserNotFoundException e) {
+                    logger.error("Error while copying content for site: " + site + " fromPath: " + fromPath +
+                            " toPath: " + toPath + " newName: ");
+                }
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while copying content for site: " + site + " fromPath: " + fromPath +
+                    " toPath: " + toPath + " newName: ");
         }
-
         return commitId;
     }
 
@@ -538,9 +569,10 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     public RepositoryItem[] getContentChildren(String site, String path) {
         // TODO: SJ: Rethink this API call for 3.1+
         final List<RepositoryItem> retItems = new ArrayList<RepositoryItem>();
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-
         try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
             RevTree tree = helper.getTreeForLastCommit(repo);
             try (TreeWalk tw = TreeWalk.forPath(repo, helper.getGitPath(path), tree)) {
 
@@ -612,7 +644,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
             } catch (IOException e) {
                 logger.error("Error while getting children for site: " + site + " path: " + path, e);
             }
-        } catch (IOException e) {
+        } catch (IOException | CryptoException e) {
             logger.error("Failed to create RevTree for site: " + site + " path: " + path, e);
         }
 
@@ -624,34 +656,39 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public VersionTO[] getContentVersionHistory(String site, String path) {
         List<VersionTO> versionHistory = new ArrayList<VersionTO>();
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+                try {
+                    ObjectId head = repo.resolve(HEAD);
+                    String gitPath = helper.getGitPath(path);
+                    try (Git git = new Git(repo)) {
+                        Iterable<RevCommit> commits = git.log().add(head).addPath(gitPath).call();
+                        Iterator<RevCommit> iterator = commits.iterator();
+                        while (iterator.hasNext()) {
+                            RevCommit revCommit = iterator.next();
+                            VersionTO versionTO = new VersionTO();
+                            versionTO.setVersionNumber(revCommit.getName());
+                            versionTO.setLastModifier(revCommit.getAuthorIdent().getName());
+                            versionTO.setLastModifiedDate(
+                                    Instant.ofEpochSecond(revCommit.getCommitTime()).atZone(UTC));
+                            versionTO.setComment(revCommit.getFullMessage());
+                            versionHistory.add(versionTO);
+                        }
 
-            try {
-                ObjectId head = repo.resolve(HEAD);
-                String gitPath = helper.getGitPath(path);
-                try (Git git = new Git(repo)) {
-                    Iterable<RevCommit> commits = git.log().add(head).addPath(gitPath).call();
-                    Iterator<RevCommit> iterator = commits.iterator();
-                    while (iterator.hasNext()) {
-                        RevCommit revCommit = iterator.next();
-                        VersionTO versionTO = new VersionTO();
-                        versionTO.setVersionNumber(revCommit.getName());
-                        versionTO.setLastModifier(revCommit.getAuthorIdent().getName());
-                        versionTO.setLastModifiedDate(
-                                Instant.ofEpochSecond(revCommit.getCommitTime()).atZone(UTC));
-                        versionTO.setComment(revCommit.getFullMessage());
-                        versionHistory.add(versionTO);
+                        git.close();
+                    } catch (IOException e) {
+                        logger.error("error while getting history for content item " + path);
                     }
-
-                    git.close();
-                } catch (IOException e) {
-                    logger.error("error while getting history for content item " + path);
+                } catch (IOException | GitAPIException e) {
+                    logger.error("Failed to create Git repo for site: " + site + " path: " + path, e);
                 }
-            } catch (IOException | GitAPIException e) {
-                logger.error("Failed to create Git repo for site: " + site + " path: " + path, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Failed to create Git repo for site: " + site + " path: " + path, e);
         }
 
         VersionTO[] toRet = new VersionTO[versionHistory.size()];
@@ -669,34 +706,39 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
         // SJ: Major revisions become git tags
         // TODO: SJ: Redesign/refactor the whole approach in 3.1+
         String toReturn = EMPTY;
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : PUBLISHED)) {
+                if (majorVersion) {
+                    Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : PUBLISHED);
+                    // Tag the repository with a date-time based version label
+                    String gitPath = helper.getGitPath(path);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : PUBLISHED)) {
-            if (majorVersion) {
-                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : PUBLISHED);
-                // Tag the repository with a date-time based version label
-                String gitPath = helper.getGitPath(path);
+                    try (Git git = new Git(repo)) {
+                        PersonIdent currentUserIdent = helper.getCurrentUserIdent();
+                        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HHmmssX");
+                        Calendar cal = Calendar.getInstance();
+                        String versionLabel = dateFormat.format(cal.getTime());
 
-                try (Git git = new Git(repo)) {
-                    PersonIdent currentUserIdent = helper.getCurrentUserIdent();
-                    DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HHmmssX");
-                    Calendar cal = Calendar.getInstance();
-                    String versionLabel = dateFormat.format(cal.getTime());
+                        TagCommand tagCommand = git.tag()
+                                .setName(versionLabel)
+                                .setMessage(comment)
+                                .setTagger(currentUserIdent);
 
-                    TagCommand tagCommand = git.tag()
-                            .setName(versionLabel)
-                            .setMessage(comment)
-                            .setTagger(currentUserIdent);
+                        tagCommand.call();
 
-                    tagCommand.call();
+                        toReturn = versionLabel;
 
-                    toReturn = versionLabel;
-
-                } catch (GitAPIException | ServiceLayerException | UserNotFoundException err) {
-                    logger.error("error creating new version for site:  " + site + " path: " + path, err);
+                    } catch (GitAPIException | ServiceLayerException | UserNotFoundException err) {
+                        logger.error("error creating new version for site:  " + site + " path: " + path, err);
+                    }
+                } else {
+                    logger.info("request to create minor revision ignored for site: " + site + " path: " + path);
                 }
-            } else {
-                logger.info("request to create minor revision ignored for site: " + site + " path: " + path);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error creating new version for site:  " + site + " path: " + path, e);
         }
 
         return toReturn;
@@ -721,10 +763,11 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public InputStream getContentVersion(String site, String path, String version) throws ContentNotFoundException {
         InputStream toReturn = null;
-
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-
         try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+
             RevTree tree = helper.getTreeForCommit(repo, version);
             if (tree != null) {
                 try (TreeWalk tw = TreeWalk.forPath(repo, helper.getGitPath(path), tree)) {
@@ -739,7 +782,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
                             " version: " + version, e);
                 }
             }
-        } catch (IOException e) {
+        } catch (IOException | CryptoException e) {
             logger.error("Failed to create RevTree for site: " + site + " path: " + path + " version: " +
                     version, e);
         }
@@ -749,117 +792,141 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public void lockItem(String site, String path) {
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            try (TreeWalk tw = new TreeWalk(repo)) {
-                RevTree tree = helper.getTreeForLastCommit(repo);
-                tw.addTree(tree); // tree ‘0’
-                tw.setRecursive(false);
-                tw.setFilter(PathFilter.create(path));
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                try (TreeWalk tw = new TreeWalk(repo)) {
+                    RevTree tree = helper.getTreeForLastCommit(repo);
+                    tw.addTree(tree); // tree ‘0’
+                    tw.setRecursive(false);
+                    tw.setFilter(PathFilter.create(path));
 
-                if (!tw.next()) {
-                    return;
+                    if (!tw.next()) {
+                        return;
+                    }
+
+                    File repoRoot = repo.getWorkTree();
+                    Paths.get(repoRoot.getPath(), tw.getPathString());
+                    File file = new File(tw.getPathString());
+                    LockFile lock = new LockFile(file);
+                    lock.lock();
+
+                    tw.close();
+
+                } catch (IOException e) {
+                    logger.error("Error while locking file for site: " + site + " path: " + path, e);
                 }
-
-                File repoRoot = repo.getWorkTree();
-                Paths.get(repoRoot.getPath(), tw.getPathString());
-                File file = new File(tw.getPathString());
-                LockFile lock = new LockFile(file);
-                lock.lock();
-
-                tw.close();
-
-            } catch (IOException e) {
-                logger.error("Error while locking file for site: " + site + " path: " + path, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while locking file for site: " + site + " path: " + path, e);
         }
     }
 
     @Override
     public void lockItemForPublishing(String site, String path) {
-        Repository repo = helper.getRepository(site, PUBLISHED);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, PUBLISHED);
 
-        synchronized (repo) {
-            try (TreeWalk tw = new TreeWalk(repo)) {
-                RevTree tree = helper.getTreeForLastCommit(repo);
-                tw.addTree(tree); // tree ‘0’
-                tw.setRecursive(false);
-                tw.setFilter(PathFilter.create(path));
+            synchronized (repo) {
+                try (TreeWalk tw = new TreeWalk(repo)) {
+                    RevTree tree = helper.getTreeForLastCommit(repo);
+                    tw.addTree(tree); // tree ‘0’
+                    tw.setRecursive(false);
+                    tw.setFilter(PathFilter.create(path));
 
-                if (!tw.next()) {
-                    return;
+                    if (!tw.next()) {
+                        return;
+                    }
+
+                    File repoRoot = repo.getWorkTree();
+                    Paths.get(repoRoot.getPath(), tw.getPathString());
+                    File file = new File(tw.getPathString());
+                    LockFile lock = new LockFile(file);
+                    lock.lock();
+
+                    tw.close();
+
+                } catch (IOException e) {
+                    logger.error("Error while locking file for site: " + site + " path: " + path, e);
                 }
-
-                File repoRoot = repo.getWorkTree();
-                Paths.get(repoRoot.getPath(), tw.getPathString());
-                File file = new File(tw.getPathString());
-                LockFile lock = new LockFile(file);
-                lock.lock();
-
-                tw.close();
-
-            } catch (IOException e) {
-                logger.error("Error while locking file for site: " + site + " path: " + path, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while locking file for site: " + site + " path: " + path, e);
         }
     }
 
     @Override
     public void unLockItem(String site, String path) {
-        Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
 
-        synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
-            try (TreeWalk tw = new TreeWalk(repo)) {
-                RevTree tree = helper.getTreeForLastCommit(repo);
-                tw.addTree(tree); // tree ‘0’
-                tw.setRecursive(false);
-                tw.setFilter(PathFilter.create(path));
+            synchronized (helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX)) {
+                try (TreeWalk tw = new TreeWalk(repo)) {
+                    RevTree tree = helper.getTreeForLastCommit(repo);
+                    tw.addTree(tree); // tree ‘0’
+                    tw.setRecursive(false);
+                    tw.setFilter(PathFilter.create(path));
 
-                if (!tw.next()) {
-                    return;
+                    if (!tw.next()) {
+                        return;
+                    }
+
+                    File repoRoot = repo.getWorkTree();
+                    Paths.get(repoRoot.getPath(), tw.getPathString());
+                    File file = new File(tw.getPathString());
+                    LockFile lock = new LockFile(file);
+                    lock.unlock();
+
+                    tw.close();
+
+                } catch (IOException e) {
+                    logger.error("Error while unlocking file for site: " + site + " path: " + path, e);
                 }
-
-                File repoRoot = repo.getWorkTree();
-                Paths.get(repoRoot.getPath(), tw.getPathString());
-                File file = new File(tw.getPathString());
-                LockFile lock = new LockFile(file);
-                lock.unlock();
-
-                tw.close();
-
-            } catch (IOException e) {
-                logger.error("Error while unlocking file for site: " + site + " path: " + path, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while unlocking file for site: " + site + " path: " + path, e);
         }
     }
 
     @Override
     public void unLockItemForPublishing(String site, String path) {
-        Repository repo = helper.getRepository(site, PUBLISHED);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repo = helper.getRepository(site, PUBLISHED);
 
-        synchronized (repo) {
-            try (TreeWalk tw = new TreeWalk(repo)) {
-                RevTree tree = helper.getTreeForLastCommit(repo);
-                tw.addTree(tree); // tree ‘0’
-                tw.setRecursive(false);
-                tw.setFilter(PathFilter.create(path));
+            synchronized (repo) {
+                try (TreeWalk tw = new TreeWalk(repo)) {
+                    RevTree tree = helper.getTreeForLastCommit(repo);
+                    tw.addTree(tree); // tree ‘0’
+                    tw.setRecursive(false);
+                    tw.setFilter(PathFilter.create(path));
 
-                if (!tw.next()) {
-                    return;
+                    if (!tw.next()) {
+                        return;
+                    }
+
+                    File repoRoot = repo.getWorkTree();
+                    Paths.get(repoRoot.getPath(), tw.getPathString());
+                    File file = new File(tw.getPathString());
+                    LockFile lock = new LockFile(file);
+                    lock.unlock();
+
+                    tw.close();
+
+                } catch (IOException e) {
+                    logger.error("Error while unlocking file for site: " + site + " path: " + path, e);
                 }
-
-                File repoRoot = repo.getWorkTree();
-                Paths.get(repoRoot.getPath(), tw.getPathString());
-                File file = new File(tw.getPathString());
-                LockFile lock = new LockFile(file);
-                lock.unlock();
-
-                tw.close();
-
-            } catch (IOException e) {
-                logger.error("Error while unlocking file for site: " + site + " path: " + path, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while unlocking file for site: " + site + " path: " + path, e);
         }
     }
 
@@ -867,9 +934,8 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
      * bootstrap the repository
      */
     public void bootstrap() throws Exception {
-        // Initialize the helper
-        helper = new GitContentRepositoryHelper(studioConfiguration, servicesConfig, userServiceInternal,
-                securityService);
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
 
         if (Boolean.parseBoolean(studioConfiguration.getProperty(BOOTSTRAP_REPO)) && helper.createGlobalRepo()) {
             // Copy the global config defaults to the global site
@@ -920,37 +986,44 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public boolean deleteSite(String site) {
         boolean toReturn;
-
-        Repository repository = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-        if (repository != null) {
-            synchronized (repository) {
-                Repository publishedRepository = helper.getRepository(site, PUBLISHED);
-                if (publishedRepository != null) {
-                    synchronized (publishedRepository) {
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repository = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+            if (repository != null) {
+                synchronized (repository) {
+                    Repository publishedRepository = helper.getRepository(site, PUBLISHED);
+                    if (publishedRepository != null) {
+                        synchronized (publishedRepository) {
+                            toReturn = helper.deleteSiteGitRepo(site);
+                        }
+                    } else {
                         toReturn = helper.deleteSiteGitRepo(site);
                     }
-                } else {
-                    toReturn = helper.deleteSiteGitRepo(site);
+                }
+            } else {
+                Path sitePath = Paths.get(studioConfiguration.getProperty(REPO_BASE_PATH),
+                        studioConfiguration.getProperty(SITES_REPOS_PATH), site);
+                try {
+                    FileUtils.deleteDirectory(sitePath.toFile());
+                    toReturn = true;
+                } catch (IOException e) {
+                    logger.error("Error while deleting site " + site, e);
+                    toReturn = false;
                 }
             }
-        } else {
-            Path sitePath = Paths.get(studioConfiguration.getProperty(REPO_BASE_PATH),
-                    studioConfiguration.getProperty(SITES_REPOS_PATH), site);
-            try {
-                FileUtils.deleteDirectory(sitePath.toFile());
-                toReturn = true;
-            } catch (IOException e) {
-                logger.error("Error while deleting site " + site, e);
-                toReturn = false;
-            }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error while deleting site " + site, e);
+            toReturn = false;
         }
-
         return toReturn;
     }
 
     @Override
     public void initialPublish(String site, String sandboxBranch, String environment, String author, String comment)
-            throws DeploymentException {
+            throws DeploymentException, CryptoException {
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
         Repository repo = helper.getRepository(site, PUBLISHED);
         String commitId = EMPTY;
 
@@ -1015,20 +1088,26 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public String getRepoLastCommitId(final String site) {
         String toReturn = EMPTY;
-        Repository repository =
-                helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-        if (repository != null) {
-            synchronized (repository) {
-                Repository repo = helper.getRepository(site, SANDBOX);
-                try {
-                    ObjectId commitId = repo.resolve(HEAD);
-                    if (commitId != null) {
-                        toReturn = commitId.getName();
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repository =
+                    helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+            if (repository != null) {
+                synchronized (repository) {
+                    Repository repo = helper.getRepository(site, SANDBOX);
+                    try {
+                        ObjectId commitId = repo.resolve(HEAD);
+                        if (commitId != null) {
+                            toReturn = commitId.getName();
+                        }
+                    } catch (IOException e) {
+                        logger.error("Error getting last commit ID for site " + site, e);
                     }
-                } catch (IOException e) {
-                    logger.error("Error getting last commit ID for site " + site, e);
                 }
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error getting last commit ID for site " + site, e);
         }
         return toReturn;
     }
@@ -1036,103 +1115,117 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     @Override
     public String getRepoFirstCommitId(final String site) {
         String toReturn = EMPTY;
-        Repository repository = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-        if (repository != null) {
-            synchronized (repository) {
-                Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
-                if (repo != null) {
-                    try (RevWalk rw = new RevWalk(repo)) {
-                        ObjectId head = repo.resolve(HEAD);
-                        if (head != null) {
-                            RevCommit root = rw.parseCommit(head);
-                            rw.sort(REVERSE);
-                            rw.markStart(root);
-                            ObjectId first = rw.next();
-                            toReturn = first.getName();
-                            logger.debug("getRepoFirstCommitId for site: " + site + " First commit ID: " + toReturn);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            Repository repository = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+            if (repository != null) {
+                synchronized (repository) {
+                    Repository repo = helper.getRepository(site, StringUtils.isEmpty(site) ? GLOBAL : SANDBOX);
+                    if (repo != null) {
+                        try (RevWalk rw = new RevWalk(repo)) {
+                            ObjectId head = repo.resolve(HEAD);
+                            if (head != null) {
+                                RevCommit root = rw.parseCommit(head);
+                                rw.sort(REVERSE);
+                                rw.markStart(root);
+                                ObjectId first = rw.next();
+                                toReturn = first.getName();
+                                logger.debug("getRepoFirstCommitId for site: " + site + " First commit ID: " + toReturn);
+                            }
+                        } catch (IOException e) {
+                            logger.error("Error getting first commit ID for site " + site, e);
                         }
-                    } catch (IOException e) {
-                        logger.error("Error getting first commit ID for site " + site, e);
                     }
                 }
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error getting first commit ID for site " + site, e);
         }
-
         return toReturn;
     }
 
     @Override
     public List<String> getEditCommitIds(String site, String path, String commitIdFrom, String commitIdTo) {
         List<String> commitIds = new ArrayList<String>();
-
-        synchronized (helper.getRepository(site, SANDBOX)) {
-            try {
-                // Get the sandbox repo, and then get a reference to the commitId we received and another for head
-                Repository repo = helper.getRepository(site, SANDBOX);
-                if (StringUtils.isEmpty(commitIdFrom)) {
-                    commitIdFrom = getRepoFirstCommitId(site);
-                }
-                if (StringUtils.isEmpty(commitIdTo)) {
-                    commitIdTo = getRepoLastCommitId(site);
-                }
-                ObjectId objCommitIdFrom = repo.resolve(commitIdFrom);
-                ObjectId objCommitIdTo = repo.resolve(commitIdTo);
-
-                try (Git git = new Git(repo)) {
-
-                    // If the commitIdFrom is the same as commitIdTo, there is nothing to calculate, otherwise,
-                    // let's do it
-                    if (!objCommitIdFrom.equals(objCommitIdTo)) {
-                        // Compare HEAD with commitId we're given
-                        // Get list of commits between commitId and HEAD in chronological order
-
-                        // Get the log of all the commits between commitId and head
-                        Iterable<RevCommit> commits = git.log()
-                                .addPath(helper.getGitPath(path))
-                                .addRange(objCommitIdFrom, objCommitIdTo)
-                                .call();
-
-                        // Reverse orders of commits
-                        Iterator<RevCommit> iterator = commits.iterator();
-                        while (iterator.hasNext()) {
-
-                            RevCommit commit = iterator.next();
-                            commitIds.add(0, commit.getId().getName());
-                        }
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(site, SANDBOX)) {
+                try {
+                    // Get the sandbox repo, and then get a reference to the commitId we received and another for head
+                    Repository repo = helper.getRepository(site, SANDBOX);
+                    if (StringUtils.isEmpty(commitIdFrom)) {
+                        commitIdFrom = getRepoFirstCommitId(site);
                     }
-                } catch (GitAPIException e) {
-                    logger.error("Error getting commit ids for site " + site + " and path " + path +
+                    if (StringUtils.isEmpty(commitIdTo)) {
+                        commitIdTo = getRepoLastCommitId(site);
+                    }
+                    ObjectId objCommitIdFrom = repo.resolve(commitIdFrom);
+                    ObjectId objCommitIdTo = repo.resolve(commitIdTo);
+
+                    try (Git git = new Git(repo)) {
+
+                        // If the commitIdFrom is the same as commitIdTo, there is nothing to calculate, otherwise,
+                        // let's do it
+                        if (!objCommitIdFrom.equals(objCommitIdTo)) {
+                            // Compare HEAD with commitId we're given
+                            // Get list of commits between commitId and HEAD in chronological order
+
+                            // Get the log of all the commits between commitId and head
+                            Iterable<RevCommit> commits = git.log()
+                                    .addPath(helper.getGitPath(path))
+                                    .addRange(objCommitIdFrom, objCommitIdTo)
+                                    .call();
+
+                            // Reverse orders of commits
+                            Iterator<RevCommit> iterator = commits.iterator();
+                            while (iterator.hasNext()) {
+
+                                RevCommit commit = iterator.next();
+                                commitIds.add(0, commit.getId().getName());
+                            }
+                        }
+                    } catch (GitAPIException e) {
+                        logger.error("Error getting commit ids for site " + site + " and path " + path +
+                                " from commit ID: " + commitIdFrom + " to commit ID: " + commitIdTo, e);
+                    }
+                } catch (IOException e) {
+                    logger.error("Error getting operations for site " + site + " and path " + path +
                             " from commit ID: " + commitIdFrom + " to commit ID: " + commitIdTo, e);
                 }
-            } catch (IOException e) {
-                logger.error("Error getting operations for site " + site + " and path " + path +
-                        " from commit ID: " + commitIdFrom + " to commit ID: " + commitIdTo, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error getting operations for site " + site + " and path " + path +
+                    " from commit ID: " + commitIdFrom + " to commit ID: " + commitIdTo, e);
         }
-
         return commitIds;
     }
 
     @Override
     public void insertFullGitLog(String siteId, int processed) {
         List<GitLog> gitLogs = new ArrayList<>();
-
-        synchronized (helper.getRepository(siteId, SANDBOX)) {
-            Repository repo = helper.getRepository(siteId, SANDBOX);
-            try (Git git = new Git(repo)) {
-                Iterable<RevCommit> logs = git.log().call();
-                for (RevCommit rev : logs) {
-                    GitLog gitLog = new GitLog();
-                    gitLog.setCommitId(rev.getId().getName());
-                    gitLog.setProcessed(processed);
-                    gitLog.setSiteId(siteId);
-                    gitLogs.add(gitLog);
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            synchronized (helper.getRepository(siteId, SANDBOX)) {
+                Repository repo = helper.getRepository(siteId, SANDBOX);
+                try (Git git = new Git(repo)) {
+                    Iterable<RevCommit> logs = git.log().call();
+                    for (RevCommit rev : logs) {
+                        GitLog gitLog = new GitLog();
+                        gitLog.setCommitId(rev.getId().getName());
+                        gitLog.setProcessed(processed);
+                        gitLog.setSiteId(siteId);
+                        gitLogs.add(gitLog);
+                    }
+                } catch (GitAPIException e) {
+                    logger.error("Error getting full git log for site " + siteId, e);
                 }
-            } catch (GitAPIException e) {
-                logger.error("Error getting full git log for site " + siteId, e);
             }
+        } catch (CryptoException e) {
+            logger.error("Unexpected error getting full git log for site " + siteId, e);
         }
-
         Map<String, Object> params = new HashMap<String, Object>();
         params.put("siteId", siteId);
         params.put("gitLogs", gitLogs);
@@ -1155,87 +1248,91 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
             throws InvalidRemoteRepositoryException, InvalidRemoteRepositoryCredentialsException,
             RemoteRepositoryNotFoundException, RemoteRepositoryNotBareException, ServiceLayerException {
         boolean toRet = true;
-        try (Repository repo = helper.getRepository(siteId, SANDBOX)) {
-            try (Git git = new Git(repo)) {
-                boolean pkauth = false;
-                final Path tempKey = Files.createTempFile(UUID.randomUUID().toString(),".tmp");
-                PushCommand pushCommand = git.push();
-                switch (authenticationType) {
-                    case NONE:
-                        logger.debug("No authentication");
-                        break;
-                    case BASIC:
-                        logger.debug("Basic authentication");
-                        pushCommand.setCredentialsProvider(new UsernamePasswordCredentialsProvider(remoteUsername,
-                                remotePassword));
-                        break;
-                    case TOKEN:
-                        logger.debug("Token based authentication");
-                        pushCommand.setCredentialsProvider(new UsernamePasswordCredentialsProvider(remoteToken, EMPTY));
-                        break;
-                    case PRIVATE_KEY:
-                        logger.debug("Private key authentication");
+        try {
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
+            try (Repository repo = helper.getRepository(siteId, SANDBOX)) {
+                try (Git git = new Git(repo)) {
+                    boolean pkauth = false;
+                    final Path tempKey = Files.createTempFile(UUID.randomUUID().toString(), ".tmp");
+                    PushCommand pushCommand = git.push();
+                    switch (authenticationType) {
+                        case NONE:
+                            logger.debug("No authentication");
+                            break;
+                        case BASIC:
+                            logger.debug("Basic authentication");
+                            pushCommand.setCredentialsProvider(new UsernamePasswordCredentialsProvider(remoteUsername,
+                                    remotePassword));
+                            break;
+                        case TOKEN:
+                            logger.debug("Token based authentication");
+                            pushCommand.setCredentialsProvider(new UsernamePasswordCredentialsProvider(remoteToken, EMPTY));
+                            break;
+                        case PRIVATE_KEY:
+                            logger.debug("Private key authentication");
 
-                        tempKey.toFile().deleteOnExit();
-                        pushCommand.setTransportConfigCallback(new TransportConfigCallback() {
-                            @Override
-                            public void configure(Transport transport) {
-                                SshTransport sshTransport = (SshTransport)transport;
-                                sshTransport.setSshSessionFactory(getSshSessionFactory(remotePrivateKey, tempKey));
+                            tempKey.toFile().deleteOnExit();
+                            pushCommand.setTransportConfigCallback(new TransportConfigCallback() {
+                                @Override
+                                public void configure(Transport transport) {
+                                    SshTransport sshTransport = (SshTransport) transport;
+                                    sshTransport.setSshSessionFactory(getSshSessionFactory(remotePrivateKey, tempKey));
+                                }
+                            });
+                            pkauth = true;
+                            break;
+                        default:
+                            throw new ServiceLayerException("Unsupported authentication type " + authenticationType);
+                    }
+
+                    logger.debug("Push site " + siteId + " to remote repository " + remoteName + "(" + remoteUrl + ")");
+                    Iterable<PushResult> result = pushCommand
+                            .setPushAll()
+                            .setRemote(remoteName)
+                            .call();
+                    if (pkauth) Files.delete(tempKey);
+
+                    logger.debug("Check push result to verify it was success");
+                    Iterator<PushResult> resultIter = result.iterator();
+                    if (resultIter.hasNext()) {
+                        PushResult pushResult = resultIter.next();
+                        Iterator<RemoteRefUpdate> remoteRefUpdateIterator = pushResult.getRemoteUpdates().iterator();
+                        if (remoteRefUpdateIterator.hasNext()) {
+                            RemoteRefUpdate update = remoteRefUpdateIterator.next();
+                            if (update.getStatus().equals(REJECTED_NONFASTFORWARD)) {
+                                logger.error("Remote repository: " + remoteName + " (" + remoteUrl +
+                                        ") is not bare repository");
+                                throw new RemoteRepositoryNotBareException("Remote repository: " + remoteName +
+                                        " (" + remoteUrl + ") is not bare repository");
                             }
-                        });
-                        pkauth = true;
-                        break;
-                    default:
-                        throw new ServiceLayerException("Unsupported authentication type " + authenticationType);
-                }
-
-                logger.debug("Push site " + siteId + " to remote repository " + remoteName + "(" + remoteUrl +")");
-                Iterable<PushResult> result = pushCommand
-                        .setPushAll()
-                        .setRemote(remoteName)
-                        .call();
-                if (pkauth) Files.delete(tempKey);
-
-                logger.debug("Check push result to verify it was success");
-                Iterator<PushResult> resultIter = result.iterator();
-                if (resultIter.hasNext()) {
-                    PushResult pushResult = resultIter.next();
-                    Iterator<RemoteRefUpdate> remoteRefUpdateIterator = pushResult.getRemoteUpdates().iterator();
-                    if (remoteRefUpdateIterator.hasNext()) {
-                        RemoteRefUpdate update = remoteRefUpdateIterator.next();
-                        if (update.getStatus().equals(REJECTED_NONFASTFORWARD)) {
-                            logger.error("Remote repository: " + remoteName + " (" + remoteUrl +
-                                    ") is not bare repository");
-                            throw new RemoteRepositoryNotBareException("Remote repository: " + remoteName +
-                                    " (" + remoteUrl + ") is not bare repository");
                         }
                     }
-                }
-            } catch (InvalidRemoteException e) {
-                logger.error("Invalid remote repository: " + remoteName + " (" + remoteUrl + ")", e);
-                throw new InvalidRemoteRepositoryException("Invalid remote repository: " + remoteName + " (" +
-                        remoteUrl + ")");
-            } catch (TransportException e) {
-                if (StringUtils.endsWithIgnoreCase(e.getMessage(), "not authorized")) {
-                    logger.error("Bad credentials or read only repository: " + remoteName +
-                            " (" + remoteUrl + ")", e);
-                    throw new InvalidRemoteRepositoryCredentialsException("Bad credentials or read only repository: " +
-                            remoteName + " (" + remoteUrl + ") for username " + remoteUsername, e);
-                } else {
-                    logger.error("Remote repository not found: " + remoteName + " (" + remoteUrl + ")", e);
-                    throw new RemoteRepositoryNotFoundException("Remote repository not found: " + remoteName + " (" +
+                } catch (InvalidRemoteException e) {
+                    logger.error("Invalid remote repository: " + remoteName + " (" + remoteUrl + ")", e);
+                    throw new InvalidRemoteRepositoryException("Invalid remote repository: " + remoteName + " (" +
                             remoteUrl + ")");
+                } catch (TransportException e) {
+                    if (StringUtils.endsWithIgnoreCase(e.getMessage(), "not authorized")) {
+                        logger.error("Bad credentials or read only repository: " + remoteName +
+                                " (" + remoteUrl + ")", e);
+                        throw new InvalidRemoteRepositoryCredentialsException("Bad credentials or read only repository: " +
+                                remoteName + " (" + remoteUrl + ") for username " + remoteUsername, e);
+                    } else {
+                        logger.error("Remote repository not found: " + remoteName + " (" + remoteUrl + ")", e);
+                        throw new RemoteRepositoryNotFoundException("Remote repository not found: " + remoteName + " (" +
+                                remoteUrl + ")");
+                    }
+                } catch (ServiceLayerException | IOException e) {
+                    logger.error("Failed to push newly created site " + siteId + " to remote repository " +
+                            remoteUrl, e);
+                    throw new ServiceLayerException(e);
                 }
-            } catch (ServiceLayerException | IOException e) {
-                logger.error("Failed to push newly created site " + siteId + " to remote repository " +
-                        remoteUrl, e);
-                throw new ServiceLayerException(e);
             }
-        } catch (GitAPIException e) {
-                logger.error("Failed to push newly created site " + siteId + " to remote repository " +
-                        remoteUrl, e);
-                toRet = false;
+        } catch (GitAPIException | CryptoException e) {
+            logger.error("Failed to push newly created site " + siteId + " to remote repository " +
+                    remoteUrl, e);
+            toRet = false;
         }
         return toRet;
     }
@@ -1248,6 +1345,8 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
         boolean isValid = false;
         try {
             logger.debug("Add remote " + remoteName + " to the sandbox repo for the site " + siteId);
+            GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                    userServiceInternal, encryptor);
             Repository repo = helper.getRepository(siteId, SANDBOX);
             try (Git git = new Git(repo)) {
 
@@ -1380,8 +1479,11 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
     }
 
     @Override
-    public List<RemoteRepositoryInfoTO> listRemote(String siteId, String sandboxBranch) throws ServiceLayerException {
+    public List<RemoteRepositoryInfoTO> listRemote(String siteId, String sandboxBranch)
+            throws ServiceLayerException, CryptoException {
         List<RemoteRepositoryInfoTO> res = new ArrayList<RemoteRepositoryInfoTO>();
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
         try (Repository repo = helper.getRepository(siteId, SANDBOX)) {
 
             try (Git git = new Git(repo)) {
@@ -1515,7 +1617,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public boolean pushToRemote(String siteId, String remoteName, String remoteBranch) throws ServiceLayerException,
-            InvalidRemoteUrlException {
+            InvalidRemoteUrlException, CryptoException {
         logger.debug("Get remote data from database for remote " + remoteName + " and site " + siteId);
         Map<String, String> params = new HashMap<String, String>();
         params.put("siteId", siteId);
@@ -1523,6 +1625,8 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
         RemoteRepository remoteRepository = remoteRepositoryDAO.getRemoteRepository(params);
 
         logger.debug("Prepare push command.");
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
         Repository repo = helper.getRepository(siteId, SANDBOX);
         try (Git git = new Git(repo)) {
             Iterable<PushResult> pushResultIterable = null;
@@ -1599,7 +1703,7 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public boolean pullFromRemote(String siteId, String remoteName, String remoteBranch) throws ServiceLayerException,
-            InvalidRemoteUrlException {
+            InvalidRemoteUrlException, CryptoException {
         logger.debug("Get remote data from database for remote " + remoteName + " and site " + siteId);
         Map<String, String> params = new HashMap<String, String>();
         params.put("siteId", siteId);
@@ -1607,6 +1711,8 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
         RemoteRepository remoteRepository = remoteRepositoryDAO.getRemoteRepository(params);
 
         logger.debug("Prepare pull command");
+        GitRepositoryHelper helper = GitRepositoryHelper.getHelper(studioConfiguration, securityService,
+                userServiceInternal, encryptor);
         Repository repo = helper.getRepository(siteId, SANDBOX);
         try (Git git = new Git(repo)) {
             PullResult pullResult = null;
@@ -1698,13 +1804,22 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public boolean isFolder(String siteId, String path) {
-        Path p = Paths.get(helper.buildRepoPath(SANDBOX, siteId).toAbsolutePath().toString(), path);
-        File file = p.toFile();
-        return file.isDirectory();
+        try {
+            GitRepositoryHelper helper =
+                    GitRepositoryHelper.getHelper(studioConfiguration, securityService, userServiceInternal, encryptor);
+            Path p = Paths.get(helper.buildRepoPath(SANDBOX, siteId).toAbsolutePath().toString(), path);
+            File file = p.toFile();
+            return file.isDirectory();
+        } catch (CryptoException e) {
+            logger.error("Unexpected error checking if path is a folder for site: " + siteId + " path: " + path);
+        }
+        return false;
     }
 
     @Override
-    public void resetStagingRepository(String siteId) throws ServiceLayerException {
+    public void resetStagingRepository(String siteId) throws ServiceLayerException, CryptoException {
+        GitRepositoryHelper helper =
+                GitRepositoryHelper.getHelper(studioConfiguration, securityService, userServiceInternal, encryptor);
         Repository repo = helper.getRepository(siteId, PUBLISHED);
         String stagingName = servicesConfig.getStagingEnvironment(siteId);
         String liveName = servicesConfig.getLiveEnvironment(siteId);
@@ -1729,15 +1844,25 @@ public class GitContentRepository implements ContentRepository, ServletContextAw
 
     @Override
     public void reloadRepository(String siteId) {
-        helper.sandboxes.remove(siteId);
-        helper.getRepository(siteId, SANDBOX);
+        try {
+            GitRepositoryHelper helper =
+                    GitRepositoryHelper.getHelper(studioConfiguration, securityService, userServiceInternal, encryptor);
+            helper.removeSandbox(siteId);
+            helper.getRepository(siteId, SANDBOX);
+        } catch (CryptoException e) {
+            logger.error("Unexpected error reloading repository for site " + siteId);
+        }
     }
 
     protected void cleanup(String siteId, GitRepositories repository) {
-        Repository sandbox = helper.getRepository(siteId, repository);
-        try (Git git = new Git(sandbox)) {
-            git.gc().call();
-        } catch (Exception e) {
+        try {
+            GitRepositoryHelper helper =
+                    GitRepositoryHelper.getHelper(studioConfiguration, securityService, userServiceInternal, encryptor);
+            Repository sandbox = helper.getRepository(siteId, repository);
+            try (Git git = new Git(sandbox)) {
+                git.gc().call();
+            }
+        }  catch (Exception e) {
             logger.warn("Error cleaning up repository for site " + siteId, e);
         }
     }
