+++ src/main/java/org/craftercms/studio/impl/v2/upgrade/pipeline/SiteRepositoryUpgradePipelineImpl.java
@@ -20,6 +20,7 @@ import java.io.IOException;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
+import org.craftercms.commons.crypto.CryptoException;
 import org.craftercms.studio.api.v1.constant.GitRepositories;
 import org.craftercms.studio.api.v1.dal.SiteFeed;
 import org.craftercms.studio.api.v1.exception.SiteNotFoundException;
@@ -30,8 +31,8 @@ import org.craftercms.studio.api.v1.service.security.SecurityService;
 import org.craftercms.studio.api.v1.service.site.SiteService;
 import org.craftercms.studio.api.v2.exception.UpgradeException;
 import org.craftercms.studio.api.v2.service.security.internal.UserServiceInternal;
+import org.craftercms.studio.api.v2.utils.GitRepositoryHelper;
 import org.craftercms.studio.api.v2.utils.StudioConfiguration;
-import org.craftercms.studio.impl.v1.repository.git.GitContentRepositoryHelper;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.lib.Ref;
@@ -102,41 +103,44 @@ public class SiteRepositoryUpgradePipelineImpl extends DefaultUpgradePipelineImp
      */
     @Override
     public void execute(final String site) throws UpgradeException {
-        GitContentRepositoryHelper helper = new GitContentRepositoryHelper(studioConfiguration, servicesConfig,
-                userServiceInternal, securityService);
+        try {
+            GitRepositoryHelper helper =
+                    GitRepositoryHelper.getHelper(studioConfiguration, securityService, userServiceInternal);
 
-        Repository repository = helper.getRepository(site, GitRepositories.SANDBOX);
-        String sandboxBranch = siteSandboxBranch;
-        if (repository != null) {
-            Git git = new Git(repository);
-            try {
-                if (!isEmpty()) {
-                    SiteFeed siteFeed = siteService.getSite(site);
-                    if (!StringUtils.isEmpty(siteFeed.getSandboxBranch())) {
-                        sandboxBranch = siteFeed.getSandboxBranch();
-                    }
-                    createTemporaryBranch(site, git);
-                    checkoutBranch(siteUpgradeBranch, git);
-                    super.execute(site);
-                    checkoutBranch(sandboxBranch, git);
-                    mergeTemporaryBranch(repository, git);
-                    deleteTemporaryBranch(git);
-
-                }
-            } catch (GitAPIException | IOException | SiteNotFoundException e) {
-                throw new UpgradeException("Error branching or merging upgrade branch for site " + site, e);
-            } finally {
-                if (!isEmpty()) {
-                    try {
+            Repository repository = helper.getRepository(site, GitRepositories.SANDBOX);
+            String sandboxBranch = siteSandboxBranch;
+            if (repository != null) {
+                Git git = new Git(repository);
+                try {
+                    if (!isEmpty()) {
+                        SiteFeed siteFeed = siteService.getSite(site);
+                        if (!StringUtils.isEmpty(siteFeed.getSandboxBranch())) {
+                            sandboxBranch = siteFeed.getSandboxBranch();
+                        }
+                        createTemporaryBranch(site, git);
+                        checkoutBranch(siteUpgradeBranch, git);
+                        super.execute(site);
                         checkoutBranch(sandboxBranch, git);
-                    } catch (GitAPIException e) {
-                        logger.error("Error cleaning up repo for site " + site, e);
-                    }
-                }
-                git.close();
-            }
-        }
+                        mergeTemporaryBranch(repository, git);
+                        deleteTemporaryBranch(git);
 
+                    }
+                } catch (GitAPIException | IOException | SiteNotFoundException e) {
+                    throw new UpgradeException("Error branching or merging upgrade branch for site " + site, e);
+                } finally {
+                    if (!isEmpty()) {
+                        try {
+                            checkoutBranch(sandboxBranch, git);
+                        } catch (GitAPIException e) {
+                            logger.error("Error cleaning up repo for site " + site, e);
+                        }
+                    }
+                    git.close();
+                }
+            }
+        } catch (CryptoException e) {
+            throw new UpgradeException("Error branching or merging upgrade branch for site " + site, e);
+        }
     }
 
     @Required
